Reorganized the blackPixelCount() function. It now...
Takes a boolean parameter to decide whether to cut off the margin or not, instead of using a global variable.
Doesn't use a global variable to store the black pixel count of an image.
Returns the integer count of black pixels, instead of the black-and-white picture.

Changed the function call to match the new parameters.